Update README.md
This is a clone website of E-commerce Website that name is "Banggood" It is 
the  first Team Project at Masai School using  HTML,CSS & JavaScript. We worked on "Banggood.in", that is a online "E-commerce website".
We had a time frame of a week . This was the first Peer Programming for all of us that given Masai Mentor.   Go on ReadMe  for more details